feat:visibility of main page improved and added styling to mainpage using motion
